+++ src/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(app)
+add_subdirectory(library)
+add_subdirectory(test_boost)
+add_subdirectory(test_google)
