+++ src/app/main.cpp
@@ -1,0 +1,10 @@
+#include "lib.h"
+
+#include <iostream>
+
+int main (int, char **)
+{
+	std::cout << "build " << buildNumber() << std::endl;
+	std::cout << "Hello, World!" << std::endl;
+	return 0;
+}
