+++ src/library/lib.cpp
@@ -1,0 +1,7 @@
+#include "lib.h"
+
+#include "version.h"
+
+int buildNumber() {
+	return PROJECT_VERSION_PATCH;
+}
