+++ src/library/lib.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int buildNumber();
