+++ src/library/version.h.in
@@ -1,0 +1,3 @@
+#pragma once
+
+#define PROJECT_VERSION_PATCH @PROJECT_VERSION_PATCH@
