+++ src/test_google/test_version.cpp
@@ -1,8 +1,9 @@
 #include <gtest/gtest.h>
+
 #include "lib.h"
 
 
-TEST(test_version, test_valid_version)
+TEST(test_google_version, test_valid_version)
 {
 	EXPECT_TRUE(buildNumber() > 0);
 }
