--- test_version.cpp
@@ -1,14 +1,0 @@
-#define BOOST_TEST_MODULE test_version
-
-#include "lib.h"
-
-#include <boost/test/unit_test.hpp>
-
-BOOST_AUTO_TEST_SUITE(test_version)
-
-BOOST_AUTO_TEST_CASE(test_valid_build)
-{
-	BOOST_CHECK(buildNumber() > 0);
-}
-
-BOOST_AUTO_TEST_SUITE_END()
