--- version.h.in
@@ -1,3 +1,0 @@
-#pragma once
-
-#define PROJECT_VERSION_PATCH @PROJECT_VERSION_PATCH@
